Added mutual information calculation, and comments
I have added the computation of mutual information via the `c3net`
library. This claims to implement the same method cited in the Rios & de
Mello paper, although I have yet to validate this claim. MI is not yet
used to partition the IMFs, as this was not described in Rios & de
Mello.

I have also added a number of comments to the existing code.